tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "colorama>=0.4.6",
    "diskcache>=5.6.3",
    "feedparser>=6.0.11",
    "firecrawl-py>=1.4.0",
    "goodreads>=0.3.2",
    "html2text>=2024.2.26",
    "markdownify>=0.13.1",
    "nltk>=3.9.1",
    "openai>=1.54.3",
    "python-dotenv>=1.0.1",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.4",
    "readabilipy>=0.2.0",
    "requests>=2.32.3",
    "spotipy>=2.24.0",
    "xmltodict>=0.14.2",
]
